extract organization code to organization-list component from governance list component.
