+++ src/app/routes/applications/governance-list/governance-list.component.html
@@ -1,37 +1,9 @@
-<!-- Filter Sub Organization -->
-<div class="d-flex align-items-center flex-wrap mb-2" *ngIf="listType === ListType.ORG && orgHierarchy?.length">
-  <div class="d-flex align-items-center my-1" *ngIf="orgHierarchy?.length">
-    <div class="filter-list-item">
-      <a href="#" (click)="resetOrgList($event)"><span class="filter-link">ALL Organizations</span></a>
-    </div>
-    <mat-icon class="color-white opacity-50">chevron_right</mat-icon>
-  </div>
-  <ng-container *ngFor="let orgItem of orgHierarchy; let i = index;">
-    <div class="d-flex align-items-center my-1" *ngIf="i !== (orgHierarchy?.length - 1)">
-      <div class="filter-list-item">
-        <img class="img-logo-list-sm mr-2" *ngIf="!orgItem?.definition?.logoUrl"
-             src="../assets/img/no-organization-image.png"/>
-        <img class="img-logo-list-sm mr-2" *ngIf="orgItem?.definition?.logoUrl" [src]="orgItem?.definition?.logoUrl"/>
-        <a href="#" (click)="resetOrgList($event, i)"><span
-            class="filter-link">{{ orgItem?.definition?.orgName }}</span></a>
-      </div>
-      <mat-icon class="color-white opacity-50">chevron_right</mat-icon>
-    </div>
-    <div class="d-flex align-items-center my-1" *ngIf="i === (orgHierarchy?.length - 1)">
-      <img class="img-logo-list-sm mr-2" *ngIf="!orgItem?.definition?.logoUrl"
-           src="../assets/img/no-organization-image.png"/>
-      <img class="img-logo-list-sm mr-2" *ngIf="orgItem?.definition?.logoUrl" [src]="orgItem?.definition?.logoUrl"/>
-      <span class="filter-link color-white opacity-50">{{ orgItem?.definition?.orgName }}</span>
-    </div>
-  </ng-container>
-</div>
-
 <!-- START: Filter-->
 <form [formGroup]="filterForm">
-  <div *ngIf="isFilterShown && listType !== ListType.ORG"
+  <div *ngIf="isFilterShown"
        class="card card-filter d-flex justify-content-end pt-2 pb-3 mb-1">
     <div class="row">
-      <div class="col-md-3" *ngIf="listType === ListType.APP || listType === ListType.ROLE">
+      <div class="col-md-3">
         <mat-label class="color-white ml-2 pl-1">Organization</mat-label>
         <mat-form-field appearance="outline" floatPlaceholder="never"
                         class="mat-input-sort line-height-filter pt-1 mr-0 mr-lg-0 mb-2 mb-lg-0">
@@ -39,7 +11,7 @@
                  maxlength="256" (input)="filter()"/>
         </mat-form-field>
       </div>
-      <div class="col-md-3" *ngIf="listType === ListType.ROLE || listType === ListType.APP">
+      <div class="col-md-3">
         <mat-label class="color-white ml-2 pl-1">Application</mat-label>
         <mat-form-field appearance="outline" floatPlaceholder="never"
                         class="mat-input-sort line-height-filter pt-1 mr-0 mr-lg-0 mb-2 mb-lg-0">
@@ -58,8 +30,7 @@
       <div class="col-md-3" *ngIf="listType === ListType.APP"></div>
       <div class="col-md-3 d-flex justify-content-center justify-content-lg-end align-items-end mb-1">
         <button mat-stroked-button class="btn btn-small mb-0 mt-2 mt-lg-0 mb-md-2 mb-lg-0" (click)="resetFilter()">Clear
-          <ng-container *ngIf="listType !== ListType.APP">Filters</ng-container>
-          <ng-container *ngIf="listType === ListType.APP">Filter</ng-container>
+          Filters
         </button>
       </div>
     </div>
@@ -75,8 +46,6 @@
       <td mat-cell *matCellDef="let element" data-label="">
         <div class="d-flex justify-content-center justify-content-md-start">
           <ng-container *ngIf="!element?.definition?.logoUrl">
-            <img class="img-logo-list" src="../assets/img/no-organization-image.png"
-                 *ngIf="ListType.ORG === listType"/>
             <img class="img-logo-list" src="../assets/img/no-application-image.png" *ngIf="ListType.APP === listType"/>
           </ng-container>
           <img class="img-logo-list" [src]="element?.definition?.logoUrl" *ngIf="element?.definition?.logoUrl"/>
@@ -87,19 +56,10 @@
     <!-- Name -->
     <ng-container matColumnDef="name">
       <th mat-header-cell *matHeaderCellDef mat-sort-header> {{ listTypeLabel }} Name</th>
-      <td mat-cell *matCellDef="let element" data-label="Name" (click)="viewSubOrgs(element)" class="cursor-pointer">
+      <td mat-cell *matCellDef="let element" data-label="Name" class="cursor-pointer">
         <div class="d-flex flex-row justify-content-between justify-content-lg-start">
-          <ng-container *ngIf="ListType.ORG === listType"> {{element?.definition?.orgName}} </ng-container>
           <ng-container *ngIf="ListType.APP === listType"> {{element?.definition?.appName}} </ng-container>
           <ng-container *ngIf="ListType.ROLE === listType"> {{element?.definition?.roleName }} </ng-container>
-          <ng-container *ngIf="element?.subOrgs && element?.subOrgs?.length">
-            <div>
-            <span class="px-2 sub-btn" matTooltipClass="multiline-tooltip" matTooltip="{{ getTooltip(element) }}"
-                  matTooltipPosition="below">
-              +{{ element?.subOrgs?.length || 0 }}
-            </span>
-            </div>
-          </ng-container>
         </div>
       </td>
     </ng-container>
@@ -110,7 +70,6 @@
         <ng-container> Role Type</ng-container>
       </th>
       <td mat-cell *matCellDef="let element" data-label="Role Type">
-        <ng-container *ngIf="element?.definition?.roleType === RoleType.ORG"> Organization</ng-container>
         <ng-container *ngIf="element?.definition?.roleType === RoleType.APP"> Application</ng-container>
         <ng-container *ngIf="element?.definition?.roleType === RoleType.CUSTOM"> Custom</ng-container>
       </td>
@@ -137,18 +96,6 @@
               <mat-icon class="btn-color-primary">preview</mat-icon>
               <span>View Details</span>
             </button>
-
-            <app-organization-actions *ngIf="listType === ListType.ORG"
-                                      [organization]="element"
-                                      (organizationCreated)="newSubOrg(element)"
-                                      (viewApps)="viewApps(ListType.ORG, element)"
-                                      (viewRoles)="viewRoles(ListType.ORG, element)"
-                                      (appCreated)="newApp(element)"
-                                      (roleCreated)="viewRoles(ListType.ORG, element)"
-                                      (edited)="edit()"
-                                      (transferred)="transferOwnership()"
-                                      (deleteConfirmed)="remove(listType, element)">
-            </app-organization-actions>
 
             <app-application-actions *ngIf="listType === ListType.APP"
                                      [application]="element"
