+++ src/app/routes/applications/organization-list/organization-list.component.html
@@ -1,0 +1,112 @@
+<!-- Filter Sub Organization -->
+<div class="d-flex align-items-center flex-wrap mb-2" *ngIf="orgHierarchy?.length">
+  <div class="d-flex align-items-center my-1">
+    <div class="filter-list-item">
+      <a href="#" (click)="resetOrgList($event)"><span class="filter-link">ALL Organizations</span></a>
+    </div>
+    <mat-icon class="color-white opacity-50">chevron_right</mat-icon>
+  </div>
+  <ng-container *ngFor="let orgItem of orgHierarchy; let i = index;">
+    <div class="d-flex align-items-center my-1" *ngIf="i !== (orgHierarchy?.length - 1)">
+      <div class="filter-list-item">
+        <img class="img-logo-list-sm mr-2" *ngIf="!orgItem?.definition?.logoUrl"
+             src="../assets/img/no-organization-image.png"/>
+        <img class="img-logo-list-sm mr-2" *ngIf="orgItem?.definition?.logoUrl" [src]="orgItem?.definition?.logoUrl"/>
+        <a href="#" (click)="resetOrgList($event, i)"><span
+            class="filter-link">{{ orgItem?.definition?.orgName }}</span></a>
+      </div>
+      <mat-icon class="color-white opacity-50">chevron_right</mat-icon>
+    </div>
+    <div class="d-flex align-items-center my-1" *ngIf="i === (orgHierarchy?.length - 1)">
+      <img class="img-logo-list-sm mr-2" *ngIf="!orgItem?.definition?.logoUrl"
+           src="../assets/img/no-organization-image.png"/>
+      <img class="img-logo-list-sm mr-2" *ngIf="orgItem?.definition?.logoUrl" [src]="orgItem?.definition?.logoUrl"/>
+      <span class="filter-link color-white opacity-50">{{ orgItem?.definition?.orgName }}</span>
+    </div>
+  </ng-container>
+</div>
+
+<!-- START TABLE -->
+<div class="card card-default mb-3 pb-3">
+  <table mat-table [dataSource]="dataSource" matSort matSortActive="name" matSortDirection="asc">
+    <!-- Logo Column -->
+    <ng-container matColumnDef="logoUrl">
+      <th mat-header-cell *matHeaderCellDef></th>
+      <td mat-cell *matCellDef="let element" data-label="">
+        <div class="d-flex justify-content-center justify-content-md-start">
+          <ng-container *ngIf="!element?.definition?.logoUrl">
+            <img class="img-logo-list" src="../assets/img/no-organization-image.png"/>
+          </ng-container>
+          <img class="img-logo-list" [src]="element?.definition?.logoUrl" *ngIf="element?.definition?.logoUrl"/>
+        </div>
+      </td>
+    </ng-container>
+
+    <!-- Name -->
+    <ng-container matColumnDef="name">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header> Organization Name</th>
+      <td mat-cell *matCellDef="let element" data-label="Name" (click)="viewSubOrgs(element)" class="cursor-pointer">
+        <div class="d-flex flex-row justify-content-between justify-content-lg-start">
+          <ng-container> {{element?.definition?.orgName}} </ng-container>
+          <ng-container *ngIf="element?.subOrgs && element?.subOrgs?.length">
+            <div>
+            <span class="px-2 sub-btn" matTooltipClass="multiline-tooltip" matTooltip="{{ getTooltip(element) }}"
+                  matTooltipPosition="below">
+              +{{ element?.subOrgs?.length || 0 }}
+            </span>
+            </div>
+          </ng-container>
+        </div>
+      </td>
+    </ng-container>
+
+    <!-- Namespace Column -->
+    <ng-container matColumnDef="namespace">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>
+        <ng-container> Organization Namespace</ng-container>
+      </th>
+      <td mat-cell *matCellDef="let element" data-label="Namespace"> {{element.namespace}} </td>
+    </ng-container>
+
+    <!-- Actions Column -->
+    <ng-container matColumnDef="actions">
+      <th mat-header-cell *matHeaderCellDef class="text-right"></th>
+      <td mat-cell *matCellDef="let element">
+        <div class="d-flex justify-content-end">
+          <button mat-icon-button [matMenuTriggerFor]="menu" aria-label="Action button list">
+            <mat-icon class="btn-color-primary">more_vert</mat-icon>
+          </button>
+          <mat-menu #menu="matMenu">
+            <button mat-menu-item class="btn-color-primary" (click)="viewDetails(element)">
+              <mat-icon class="btn-color-primary">preview</mat-icon>
+              <span>View Details</span>
+            </button>
+
+            <app-organization-actions [organization]="element"
+                                      (organizationCreated)="newSubOrg(element)"
+                                      (viewApps)="viewApps(element)"
+                                      (viewRoles)="viewRoles(element)"
+                                      (appCreated)="newApp(element)"
+                                      (roleCreated)="viewRoles(element)"
+                                      (edited)="edit()"
+                                      (transferred)="transferOwnership()"
+                                      (deleteConfirmed)="remove(listType, element)">
+            </app-organization-actions>
+          </mat-menu>
+        </div>
+      </td>
+    </ng-container>
+
+    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+  </table>
+  <!-- Shows when table is empty -->
+  <div class="d-flex justify-content-center align-items-center mt-4 mb-3" *ngIf="!dataSource?.data?.length">
+    <div class="d-flex align-items-center no-record color-white">
+      <mat-icon class="mr-2">info</mat-icon>
+      <i>No Record Found</i>
+    </div>
+  </div>
+  <!-- <mat-paginator [pageSizeOptions]="[5, 10, 20]" [hidePageSize]="true" showFirstLastButtons>
+  </mat-paginator> -->
+</div>
