+++ src/app/routes/applications/organization-list/organization-list.component.scss
@@ -1,0 +1,4 @@
+.mat-menu-item[disabled], .mat-menu-item[disabled]::after,
+.mat-menu-item[disabled] .mat-icon-no-color, .mat-menu-item-submenu-trigger[disabled]::after {
+  color: rgba(255, 255, 255, 0.3) !important;
+}
