+++ src/app/routes/dashboard/card/card.component.html
@@ -1,5 +1,6 @@
 <div class="card d-flex justify-content-center align-items-center py-3 mb-3"
      [class.card-icon-disabled]="comingSoon"
+     [attr.data-qa-id]="title"
      [class.card-icon]="!comingSoon">
   <div *ngIf="comingSoon" class="position-fixed d-flex align-items-center text-center">
     <h3 class="color-white opacity-50">Coming Soon</h3>
